@@ -96,16 +101,17 @@
     "lottie-web": "5.12.2",
     "minimalistic-crypto-utils": "1.0.1",
     "moment": "2.30.1",
-    "node-polyfill-webpack-plugin": "2.0.1",
     "package-json": "7.0.0",
     "patch-package": "6.5.1",
     "platform": "1.3.6",
+    "postcss": "8.4.35",
     "qrcode-with-logos": "1.0.5",
     "register-service-worker": "1.7.2",
     "scrollmagic": "2.0.8",
     "stream": "0.0.2",
-    "subscriptions-transport-ws": "0.11.0",
-    "vue": "2.6.14",
+    "vite": "5.1.5",
+    "vite-plugin-node-polyfills": "^0.21.0",
+    "vue": "2.7.0",
     "vue-apollo": "3.1.2",
     "vue-chartjs": "3.5.0",
     "vue-lazyload": "1.3.4",
@@ -144,10 +144,9 @@
     "codecov": "3.8.3",
     "commitlint": "18.6.1",
     "copy-webpack-plugin": "6.3.2",
-    "eslint": "7.32.0",
-    "eslint-plugin-prettier": "4.2.1",
+    "eslint": "8.57.0",
     "eslint-plugin-security": "1.7.1",
-    "eslint-plugin-vue": "9.0.1",
+    "eslint-plugin-vue": "9.22.0",
     "ethereumjs-util": "7.1.5",
     "ethereumjs-wallet": "1.0.2",
     "geckodriver": "3.2.0",
@@ -195,8 +193,8 @@
     "xss": "1.0.14"
   },
   "engines": {
-    "node": ">= 10.0.0 < 12.3.0",
-    "npm": ">= 6.0.0"
+    "node": ">= 17",
+    "npm": ">= 8.0.0"
   },
   "resolutions": {
     "@keepkey/device-protocol": "7.8.1",
