@@ -1,5 +1,5 @@
 const lokalise = (translation, string) => {
-  if (process.env.VUE_APP_MODE === 'translate') {
+  if (import.meta.env.VITE_APP_MODE === 'translate') {
     return `{.${string}.}`;
   }
   return translation;
