@@ -1,5 +1,5 @@
 // eslint-disable-next-line
-if (BUILD === 'offline') {
+if (import.meta.env.BUILD === 'offline') {
   import('@/main/offlineMain.js');
 } else {
   import('@/main/liveMain.js');
